add item touch helper to transaction adapter

+++ app/src/main/java/com/example/currencyconverter/data/database/CurrencyTransaction.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class CurrencyTransaction(
     @PrimaryKey(autoGenerate = true) var id: Int = 0,
     var name: String,
-    var value: Int
+    var value: Double
 )
